Cleaned-up VA test cases.
